Patch for Brent Nicolle.
Unit tests of code in the Core plugin should now be placed in the core.tests plugin. 
(I did not delete test from the ui.tests plugin, but that's an idea whose time is coming soon.) 

Resources and property files for unit tests are now in a separate directory from the root. 

The testlauncher may experience some turbulence, but out of the box tests work as before.