Patch for Andrew Niefer
partial fix for bug 43664 - Search cannot report matches outside of 
workspace.

This enables non-ui reporting of matches outside the workspace.  So 
clients of search, like code assist will get external results.

We still need a resource to report a match to the search view so no 
external matches in the UI.

core:
Modify Matchlocator to not try and create a link if we have no resource, 
instead just use the path

core.tests:
added testNoResourceSearching() to OtherPatternTests

ui:
modify CSearchResultCollector to accept matches without resources, but not 
attempt to report
them in the UI.
* src/org/eclipse/cdt/internal/ui/search/CSearchResultCollector.java