Patch for Sean Evoy: Two of the fixes, 43614 and 43756, involved changing property files only, which validates the extra work of externalizing strings from the start!
For 43616, I simply took the advice of the bug reporter and added the '-' 
in front of the RM macro in the clean target and the include directives in 
the makefile.

The largest part of the fix involves 43220. Until just now, this was a 
critical bug in bugzilla, so I addressed it. It has just been downgraded 
to an enhancement request. There is now a new entry widget in the linker 
options for user objects. The makefile will simply add these to the final 
build target's command. Most of the work was done in the plugin file and 
the build model to handle the new "type" of option.