Fix for 83556 -- The solution involved several fixes. First, there were use cases that failed because the MBS does not actually add a new config until the OK button is hit. The user can delete and restore this new config as often as they want until then. They might also delete existing configs. The solution to this problem was to store the deleted "new" configs in a separate data structure.
Another problem was the name identity check. The name is now checked in a case
insensitive way. The name is also checked for whitespaces and invalid
directory name characters.

Finally, the error dialog was removed and the error message displayed in a
status line with the OK button will be disabled until the user
corrects the problem.