Bug 226931 Support for Restart button. The steps to restarting the inferior are the following: 1- Create a new PTY and tell GDB to use it 2- Create a new MIInferiorProcess object which uses the new PTY 2.5- Have the CLIEventProcessor use the new MIInferiorProcess 3- Restart the inferior using -exec-run 4- Remove the previous inferior Process from the launch 5- Add the new inferior Process to the launch (which will trigger the use of the new PTY streams)
This change supports the Restart function, including the above steps to perform the proper cleanup.  The code to start the inferior has been extracted from the FinalLaunchSequence and put in GDBControl to allow sharing between start and restart.  Also, the code to create the CLI and inferior process objects has been extracted from the GdbLaunchDelegate and put in GDBControl to to allow sharing between start and restart.

There only interface change that is not in a provisional interface is the
addition of resetInferior() to CLIEventprocessor which is backwards compatible.