[252782] Adapt the launch sequence to the latest non-stop syntax.
"set breakpoint apply global" does not exist anymore.
"set breakpoint always-inserted" is no longer needed because GDB does it
automatically for non-stop.
"maint set remote-async 1" and "maint set linux-async 1" are replaced by a
simpler "set target-async on"

Also, I changed the version of the GDB that supports multi-process.  This GDB
is not actually available publicly yet, but it works for the preliminary
release I have.  Once GDB 7 is out, we will change the version again.