Update patch set 1
Patch Set 1: Verified; IP review completed

(3 inline comments)

Thanks Pawel for working on this!  

Note that the behavior of reverse buttons is not behaving nicely with the 4.2 platform.  So, for this review, I used 3.8.  Outside of this review, I will do some tests with 4.2 to report the issues to platform UI.

There are a couple of comments within the files, and here are a couple of global ones:

- Button order of global debug toolbar is very different that debug view toolbar.  Would it be possible to make the global toolbar order the same as the debug view toolbar?

- If I enable both debug toolbars, when pressing one of the two reverse toggle button, the other one's state does not update until the DV selection is updated.

- When enabling/disabling the visibility of the DV debug toolbar, the reverse button are not shown/don't disappear until the DV selection is updated.  This is similar to when the user presses the reverse toogle, which was fixed in bug 290572 by using the IEvaluationService in ReverseToggleCommandHandler.postExecute()


P.S. I'm marking at +1 verified since it compiles :) (and Hudson is not hooked to Gerrit for CDT).  And I'm marking at +1 IP clean, obviously, but I'll leave the code review to 0 due to the comments.  Thanks again!

Patch-set: 1