Update patch set 3
Patch Set 3: (1 inline comment)

Guys you should not do any formatting changes along with the changes submitted for review (and in a commit as well). That makes more difficult to figure out what was changed. If you feel you must do it do it as a separate commit. Especially if somebody else maintains the code.

Patch-set: 3