Update patch set 1
Patch Set 1: Fails; Unclean IP, do not check in

(6 inline comments)

Hi Anton.  Awesome feature!  This is really cool.

Don't worry about the two -1s, let me explain.  The missing @since tags cause 'compiler' errors, so technically, it does not compile.

Also, you must add a copyright line with your name, the company you work for, and a small description at the top of each file you modified.  That is why the IP Clean is -1.  You can mimic this, taken from MIVariableManager:

 *     Jens Elmenthaler (Verigy) - Added Full GDB pretty-printing support (bug 302121)

The rest of the comments are in-line and should be very easy to fix.

Did you think about JUnit tests?

Thanks a lot for the contribution, I will be able to commit it for Juno, once you addressed the minor points.

Patch-set: 1