Update patch set 2
Patch Set 2:

Hi Alex, I tested the new mode "Run on file open" (only this mode selected) with the Assignment to itself checker and it worked. However, I tried "Run on file save" (only this mode selected) with the same checker and it didn't run. Does that work for you? Is this mode expected to work with any checker?

Patch-set: 2