Update patch set 1
Patch Set 1: Verified; Looks good to me, approved; IP review completed

(7 inline comments)

Nice fix.

I put minor comments (such as missing javadoc and some point I was clear about).

Surprisingly though, I've seen that once in a while (1 in 5 or 10 times), the launchesTerminated() method in WaitForTerminationJob does not get called at all, and the launch is not removed as expected when doing "Terminate and remove". Then, eventually, the 1 minute timeout hits in WaitForTerminationJob.

I don't understand why this is happening. Maybe a bug in platform? From what I see, the WaitForTerminationJob is properly started, but the launchesTerminated() is not called at all, in those cases.

I don't think you should delay committing this fix because of this, as things are definitely better than before.  But a bug should be opened (or the problem investigated further).

Thanks

Patch-set: 1