Update patch set 1
Patch Set 1: No score

I just had a thought that could make the solution simpler.  I didn't actually try it though.

How about having the WaitForTerminationJob register a SessionEndedListener with the DsfSession instead of using the launchesTerminated() callback?  That way, the new callback would only happen when the launch is fully terminated as we need.  That would also remove the need for the new GdbLaunch.isShutdown() API.  And hopefully, it would remove the apparent bug that I mentioned where launchesTerminated() is not being called.

What do you think?

Patch-set: 1