Update patch set 2
Patch Set 2:

Hi Alex. A problem I see with this change is that using "on file save" runs when the file is changed externally, I think it should be considered a file open when it reloads the file. What do you think? Perhaps CDT needs the equivalent of IPostSaveListener (in JDT) or a lite version of it to achieve proper "on file save".

Patch-set: 2