Update patch set 2
Patch Set 2:

It is possible to do it without using a job. In this case we need to store all requests associated with a session in DsfTerminateHandler. Requests (including those for multip-rocess sessions) would be kept until the associated session is ended. The disadvantage of this approach is that some of the requests could be kept for a long time, otherwise we need to implement a timeout which more complicated and would require a job.
I am checking the patch in. If you still have concerns let me know and I will change the code.

Patch-set: 2
Label: SUBM=+1