Bug 365718: Support tree option style in managed build options
Managed build options in tools provide several styles
for option type, for example: string, boolean, libPaths, enumerated, ...
The enumerated type allows the user to select from a pre-defined list of
values.

A usability problem might happen if the list of possible values is Huge,
in this case the simple drop-box wouldn't be sufficient for the user to
properly and easily select the needed option. In most of those cases the
huge list of options can be organized in some hierarchy that allows the
user to easily pick the needed option.

The attached patch attempts to support this by doing the following:

1- Adding support of a new option valueType "tree"
    - Adding schema for treeOption
    - Adding support for tree type similar to enumerated option type
2- Provide UI to easily select items from this tree
    - Provide a field editor with a text box and browse button
    - Provide popup tree picker with filtered search
    - Each item in the tree can optionally provide an icon, and specify
its relative order among its peers.

Updates:
- Expose TreeSelectionDialog and TreeRoot to allow usage in other
contexts
- Remove extra whitespaces.
- Support Double-click to select
- Added unit test

Change-Id: I3e7e8c6e3e9878f7c6dadeb149326b375cd3631d
Reviewed-on: https://git.eclipse.org/r/5558
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>
Tested-by: Doug Schaefer <dschaefer@qnx.com>