Update patch set 3
Patch Set 3:

I added test for the core part of the tree option trying to cover as much as possible of the new/modified APIs.
Please let me know if this isn't enough.
Part of the Patch is in ui, so I didn't add test for that, is Jubula used in cdt, are there ui tests?

Patch-set: 3