Update patch set 3
Patch Set 3: Verified; Looks good to me, approved; IP review completed

Looks good. I think there are some usability issues with how you select items in the tree. It's not obvious that you can't select parent nodes, for example. But since this won't be visible in the open CDT product, we have time to work on those. Also, the option in the test definition needs to be in a category so that it's visible in the UI to play with. Submitting.

Patch-set: 3
Label: SUBM=+1