Update patch set 1
Patch Set 1:

This is a first suggestion how User Defined Literals could be implemented syntactically. The goal is to avoid receiving error markers in the editor when working with C++11 code, specifically user defined literals, although they are not completely covered semantically.

In this approach I've added four new kinds of tokens, a user defined version of integer, floating, string and char. They could be further split up or consolidated to one USER_DEFINED_LITERAL token.

For avoiding to break existing functionality that conflicts with the definition of user defined literals for C++ in N2765, the recognition currently requires a user defined literal to have a suffix starting with an underscore.
For not conflicting with existing features (e.g 0b1 which could be recognized as an integer literal in binary form or a user defined literal) an additional lexer option for user defined literal support could be added.

Patch-set: 1