Update patch set 2
Patch Set 2: Verified; Looks good to me, approved; IP review completed

I've updated the change to make sure that if the GDB binary is not found, the tests will be ignored.

I will commit this now.  What should happen on Hudson is that we'll see 326 new tests being ignored.

Next step is to add GDB on the Hudson Linux machine.

Patch-set: 2
Label: SUBM=+1