Update patch set 1
Patch Set 1: (2 inline comments)

This is a nice improvement on the current situation.

Before going deeper in the details let me ask if it wouldn't be better to handle the errors in the caller of IRunControl.resume() (for example)?  It seems more appropriate to report such an error in the UI code which triggered the failing operation.  For example, DsfResumeCommand would handle the failure and show the dialog.  I don't think the service should be showing errors to the user; instead it should report the error as it does now with a failed status, and the different callers should decide what to do with it.

What do you think?

Patch-set: 1