Update patch set 1
Patch Set 1:

I agree that's the right way to do it. This would requires to go through all commands and add error handlers. Some commands like 'Run To Line', 'Resume At Line' and 'Move To Line' already have error handles. 
The problem is we don't exactly whether there are commands except 'Resume' and 'Step' that are affected by this problem. Should we fix it for 'resume' and 'step' and wait for bug reports for other cases?

Patch-set: 1