Update patch set 1
Patch Set 1:

This is a first suggestion to implement C++11 Explicit Virtual Overrides.

Parser: New keywords 'override' and 'final' have been introduces. As they are keywords only in specific contexts I did not add them as tokens to IToken. The parser looks for an IDENTIFIER and checks whether its image matches the new keywords.

AST: The ICPPASTFunctionDeclarator and the ICPPASTCompositeTypeSpecifier have been extended to allow setting and querying the new properties.

Bindings: ICPPMethod and ICPPClassType provide methods to access the new properties as well.

AST Writer: The DeclaratorWriter and DeclSpecWriter have been extended to write the new properties to the scribe for the corresponding AST nodes.

Patch-set: 1