Update patch set 3
Patch Set 3: (14 inline comments)

I've made the changes requested, including the changes to GdbCommandTimeoutManager. I'll submit them in patch set 4. The only question I have right now is whether we're going to do automatic wrapping at 100 characters for the file write in trace or just for the sysout. The patch set is written to not do auto wrapping on the file write, but it's a quick change if we want to go the other way.

Patch-set: 3