Update patch set 7
Patch Set 7: Verified; IP review completed

Thanks Jason, this one's the one.

I'm not marking as approved yet because we have to wait until Juno is released before I can commit.  Once Juno is released and the master branch is available again, I will commit this.  Don't be shy to ping me in case I forget to get back to this.

Patch-set: 7