Update patch set 2
Patch Set 2:

Memory monitor is the same as memory block. 
In my new patch the memory monitors are saved on when a ContainerExitedEvent is fired instead of being saved on shutdown. The monitors are also removed from the view which will fix https://bugs.eclipse.org/bugs/show_bug.cgi?id=383683.
To add a full multi-process support we need to save the process information (the name of the executable?) with the address expression.

Patch-set: 2