Update patch set 1
Patch Set 1:

Marc, could you please review this patch? The affected code is in DSF but I don't really know who else is interested in it. Maybe Randy?
The fix is simple. The memory view tracks the content of the memory blocks by setting the 'HISTORY_KNOWN' flag for each byte. For some reason 'dsfMemoryBlock' sets the flag only is the byte value has changed. If the unit size is greater than 1 and only one byte in the unit has changed the remaining bytes in the unit don't have the 'HISTORY_KNOWN' flag set. As a result the entire unit is not regarded as being tracked.

Patch-set: 1