Update patch set 1
Patch Set 1:

> Looks to me like DSF should have set the HistoryKnown flag all the time from the start.

I agree, all bytes read from the backend that are not '??' should set the 'HISTORY_KNOWN' flag.

> I wonder why this problem was not visible before?

The code has been basically the same since the DSF plugins were moved to CDT. The only explanation is the most of the DSF clients use the traditional memory rendering with the memory view.

Patch-set: 1