Bug 385216 - Convert org.eclipse.cdt.make.core to Platform Debug Trace
I've made the necessary conversions. Since all debugging was running
through the TraceUtil class anyway, I thought I'd maintain that same API
and just reroute that class to use the dynamic options, so that messages
stay consistent. We're completely replacing the LogWriter class with the
debug options class, but I'm not removing the class so that I don't
break API.

Change-Id: I286604e69a892f8031498168e9316dae13beb963