Bug 385238 - Convert org.eclipse.cdt.managedbuilder.core to Platform Debug Trace
I've added dynamic debugging to the managed builder plugin. I've kept
the DbgUtil class and redirected the trace to a new debug options class.
I think eventually, we could migrate DbgUtil to the options class, but
we would need to be able to deal with the prefixes and postfixes that
are added to the debug strings to keep consistency.


Change-Id: I2d20aace001b7b88d5a67de81691a7bb6f2398f2