Update patch set 4
Patch Set 4: (9 inline comments)

I reviewed the new patch based on the previous comments and here are a couple more minor things to tweak.  The comments are on patch 4 because that is where I started to review, but they are all applicable to patch 5

I still have to try out the patch and review a bit more before approving.

Patch-set: 4