Update patch set 1
Patch Set 1:

I just realized that you made this commit on top of the "Trace Control View Enhancements" commit.  So, Gerrit thinks that this change requires the other (see the "Dependencies" section in the Gerrit review.

Can you cherry-pick this commit directly onto master and submit it again to avoid confusion when I commit it first?

To make sure you stay on the same Gerrit review, you should use the same "Change Id" value in your new commit.  So the full procedure would be:
1- check out master
2- cherry pick this commit
3- do an empty commit to update the commit comment and use changeId: Ibc18201c3e0c34d4c3524d463cd6d850a3f2dead
4- push to gerrit

Thanks

Patch-set: 1