Update patch set 3
Patch Set 3: Verified; I would prefer that you didn't submit this; IP review completed

(8 inline comments)

I noticed that this change breaks some JUnit tests.  Namely org.eclipse.cdt.tests.dsf.gdb.tests.LaunchConfigurationAndRestartTest

The reason is that the case of a \n is not handled anymore.  It may sound far fetched, but it actually is not because users can easily put arguments in two different lines in the launch Arguments tab.  I recall there was a bug about this but I cannot find it.

Also, I'm bothered by the change in DebugNewProcessSequence though.  We are changing the behavior of MIGDBSetArgs.  From what I understand, MIGDBSetArgs should be passed a single string of arguments, to give it "as-is" to GDB.  We should not be using an array of arguments.  Therefore, I think we should add a new

MIGDBSetArgs(IMIContainerDMContext, String)

and corresponding

public ICommand<MIInfo> createMIGDBSetArgs(IMIContainerDMContext dmc, String arguments)

we should also deprecate
MIGDBSetArgs(IMIContainerDMContext, String[])
since it should not be used anymore, and the one in CommandFactory.

In org.eclipse.cdt.debug.gdbjtag.core:
GDBJtagDSFFinalLaunchSequence should be updated not to use the newly deprecated createMIGDBSetArgs(IMIContainerDMContext, String[])

Patch-set: 3