Replacement of the CPPNameCollector with BindingAssertionHelper in Template Alias tests.
Sergey suggested to make the Template Alias tests more readable by
replacing the CPPNameCollector with BindingAssertionHelper. I did this
where it was possible. I.e it was still required in tests asserting an
instance count of a binding. But to acquire those references the
BindingAssertionHelper is used as well. For comparison of those bindings
I have added a method for creating a CPPNameCollector with a given AST.

Change-Id: I19b11cbabe5c4c0fa57965d6f97e0db5dfa5575a
Reviewed-on: https://git.eclipse.org/r/9020
Reviewed-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>
IP-Clean: Sergey Prigogin <eclipse.sprigogin@gmail.com>
Tested-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>