Update patch set 1
Patch Set 1: Verified

The change makes sense. I'm puzzled we didn't discover this previously, but maybe it's because it only shows up when trying to represent more than 2 cores on a cpu. Maybe we only tested it live on boxes with dual-core cpus? 

One suggestion: adding the null check on the model is good, but add a check for zero on the division, e.g.:

int ncores = m_cores.size() / ((ncpus == 0) ? 1 : ncpus);

Patch-set: 1