Test moving the scheduling of the RSEInitJob to the RSEUIPlugin
The RSEInitJob loads profiles which instantiates subsystem configurations.
If these subsystem configurations are located in plugins that depend on
UI plugins then we run the risk of activating these plugins before
the platform is ready resulting in load errors of various kinds.

There appears to be a change in job scheduling timing that makes this
error more apparent.