Update patch set 4
Patch Set 4:

Removed redundant parentheses.

The patch fixes a problem with remote include files which are not being correctly recognized in the current version of CDT. This is addressed in InternalFileContentProvider#getInclusionExists(). The change in InternalParserUtil was made because it was assumed that this code could be used, but it turns out that this is already handled in InternalParserUtil#createExternalFileContent().

Patch-set: 4