Update patch set 4
Patch Set 4:

Simply to allow the call to store.fetchInfo().exists() to be cancelable. I could remove the job if you wish, but if there as a network issue then this call could potentially take forever. It's called by the CPreprocessor when processing include files, so I'm not sure what the implications of blocking this operation would be.

Patch-set: 4