Update patch set 1
Patch Set 1: Do not submit

I understand what you're shooting for here, but there's
a more correct way to achieve it, by using the dispose()
methods rather than trying to replace them.

The "dispose()" method serves two purposes:
(a) to allow an instance to be explicitly discarded,
    releasing any associated resources
(b) to allow other classes that use the class to
    dispose of it as part of their own dispose() call
(As a side note, the dispose() method should be written
so that it tolerates being called more than once.)

dispose() is not intended to act like a destructor (and in
fact cannot), but more like a "release" or "free",
allowing the user of a class to explicitly dispose of it
at a chosen point rather than simply dropping the reference
and letting the garbage collector deal with it at some
arbitrary time in the future.

A parent class may need to dispose() of its
children in some specific order, so in general one should
let the "owner" of an object reference manage its disposal.

The proper way to resolve this issue is to have
the outermost owning class (the view) detect when
it has closed (perhaps by using a disposeListener in
that class's code, as you're doing), and have it invoke
its dispose() method. That dispose() method then invokes
the dispose()s of any parent/referenced classes, so that
the intended dispose() order of the class hierarchy
is maintained.

(Side Note: it looks like the change in MulticoreVisualizerCanvas should be in its own bugfix,
since it appears to be unrelated to this change.)

Patch-set: 1