Update patch set 5
Patch Set 5:

I tested the cancel operation and it doesn't seem to work with Jsch, and the connection just eventually times out. In addition, it looks like the existing code in InternalParserUtil#createExternalFileContent() doesn't provide a progress monitor to the toLocalFile() method which is likely to take longer anyway. In light of this, I've revised the patch and removed the use of Job. There's still likely to be problems with connections hanging, etc., but at least remote include files will work.

Patch-set: 5