Update patch set 3
Patch Set 3:

Hi again,

I have rebased the patch to include bug 399281.  Fixed a race condition related to the disposal of the view in the load meter timer and cleaned-up some dead code in MulticoreVisualizerCanvas: the load meters can only be updated, not created in canvas, so removed creation code.

Regards,

Marc

Patch-set: 3