Bug 400020: Prototype for the tagger extension point
This prototype is further developed than what is currently described
in the bug (I'll update the bug shortly) it now allows tagging
of non-persistent IBinding's.  These tags are not persisted.

One biggest problem in this prototype is the algorithm for finding
the project that initiated the indexing operation.  This happens in
TaggerDescriptor.  The goal is to find the list of project natures
so that the extensions enablement expression can be evaluated.  I
haven't found a better way yet.

Change-Id: I82d6f7d9675ce908bd789f42d93c5b8a5f4418c6