Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

(79 inline comments)

Nice work!  The feature looks good and integrates well.  It will be great addition to the Visualizer.

I wrote some in-line comments but there is nothing major in them.  There is only one issue I am a bit concerned about, which is:  

1- when I unselect the debug session in the debug view, or select another session, the cpu timer stops.  Let's discuss if it should or not.

This brings up a tricky point.  The visualizers, up to now, had to focus on getting information when the view was being displayed.  If we want to have a concept of HighWaterMark, it implies gathering information even if the view is not being displayed currently.  To properly support the highwater mark, I believe we need to keep monitoring the target as long as the debug session is active and the visualizer view is not destroyed.

That means that if we select another debug session or un-select the current one, we somehow need to keep monitoring the load to properly manage the high water mark.  If so, then we need to figure out who need to keep polling, the view or the service.  If it is the view, then we need to continue dealing with sessions that are on-going.  The class OSResourcesView does this with SessionOSData using fSessionDataCache.  The other option is to do the polling in the service.  Since we are already setup to have the view poll, and that OSResourcesView also does that, I suggest we try mimicking what OSResourcesView does.

Below are general comment when I was testing:

2- how difficult would it be to draw the cores bigger if the load meters are not shown since we don't have to keep space free on the margin for the load of the CPU?

3- On my machine I cannot see what is selected for the load refresh speed, there are no checkmarks or anything

4- When I disable the load meters, the view does not refresh until I select something in the Debug View, so the load meter remain there
  When I enable, they only appear after the first timeout, which can seem a bit long.

5- Do we really want to be able to set the refresh speed even if the Load meters are disabled?  It might simply things if that sub-sub-menu was disabled if the meters are disabled

6- Can we handle the case where getting the load returns an error in a more user-friendly way?  Maybe a pop up using GdbStatusHandler and disabling the enable command?

7- Say we cannot get the load for 5 times in a row, maybe we should disable the feature, and require the user to manually re-enable it? Or else we'll keep sending request to GDB I believe

8- I think having the load meter actions in the view menu may be more appropriate than the context menu.  What do you think?

9- when I kill the visualizer view, and re-open it, it does not remember that I had enabled the load meters.  I think it should, no?
  We'll need to store this setting in the preference store.

Patch-set: 4