Update patch set 1
Patch Set 1:

I don't think that the binding in an EvalBinding is the right thing to derive the scope from. The "point of definition" we are looking for is the point of definition of the template containing the dependent expression, not of a binding that the dependent expression might mention.

Consider the following example:


template <typename T>
struct A
{
    static T x;
};

struct S{};

namespace N
{
    S operator++(S);

    template <class T>
    struct impl
    {
        typedef decltype(++(A<T>::x)) type;
    };
}

void test()
{
    typedef N::impl<S>::type type;
}


Here, the EvalBinding being wrapped wrapped by the EvalUnary is A<T>::x. The scope that we derive from that, the class scope of A<T>, is the wrong scope to look up the operator++. We want the scope of the binding impl<T>::type, which is what contains the dependent expression. Since this scope is external to the dependent expression, I don't think it can be derived from the contents of the dependent expression.

Patch-set: 1