Update patch set 1
Patch Set 1:

That should be possible. Evaluations are typically created by IASTExpression.getEvaluation(). We can walk the AST up from the IASTExpression to see if we encounter an ICPPASTTemplateDeclaration. If so, we can get the declaration's name and resolve it to get the binding for the template. If we don't see an ICPPASTTemplateDeclaration, we can pass null for the binding (similar to how in this patch, null is passed as the point of definition if we're not in a template context).

Does that seem reasonable? I can give it a try.

Evaluations would have to persist the binding in the PDOM. I guess that's not a problem, other than that it makes the index larger?

Patch-set: 1