Update patch set 1
Patch Set 1:

I agree with Sergey that the code is easier to understand if we store the point of definition with the dependent names, i.e. in the Evaluation objects.

The recursion could be overcome be determining the point of definition lazy, i.e. when the evaluation object needs to be serialized, or when it is instantiated.

Patch-set: 1