improvements to the prototype implementation
The tag functionality has been extended to apply to all types of
IBindings (instead of just the PDOMBindings).  Tags for non-persistent
bindings are created on demand.  Creation of a new PDOMBinding is
one place that the original binding's tags are "demanded".

Also removes hacks that were used to find the indexing project.
Instead of operating on only an IBinding, the taggers now receive the
binding as well as the originating IASTName.  This provides a direct
lookup of the originating translation unit and project.

The Taggables are now accessible through the ITaggableService of
CCorePlugin.

In-memory tags and taggables are not cached.  I've done some timing
tests using my sample implementation and found no measurable difference.
The full log lines look like:

	!MESSAGE Indexed 'simple-01' (2 sources, 184 headers) in xxx sec:
21,550 declarations; 35,394 references; 0 unresolved inclusions; 1
syntax errors; 0 unresolved names (0.00%)

I did 5 tests using the current master (no tagging-related code), the
times were:
	18.86 sec
	 9.17 sec
	 5.91 sec
	 4.79 sec
	 4.83 sec

And then I ran the same sequence of tests using the code in this
commit:
	18.73 sec
	 9.39 sec
	 6.50 sec
	 4.78 sec
	 5.27 sec

If performance does become a problem, then caching could be introduced
with a new implementation of ITaggableService.  The two problems are
finding a key other than the identity of the IBinding (since IBindings
are re-created often) and properly evicting old entries when the
binding is no longer valid.

Change-Id: I82d6f7d9675ce908bd789f42d93c5b8a5f4418c6