Bug 400020: remove stale tags and ensure proper DB locking
The previous implementation did not remove stale tags from the
index when they were no longer applicable.  This updates the interface
to provide a place for tags to be removed and updates the PDOM
implementation to remove those tags.  There is no need to update the
in-memory implementation since tags are always recreated on demand.

----
The implementation allows existing tags to be updated in-place when
possible.  In the common case, the tag can be updated in place.  The
binding will be stored at the same offset, the same taggers are usually
still applicable, and the size of the tag should rarely change.  The new
data is just copied onto into the existing buffer.

The next most common case is when the source code has changed in a way
that means the tag is either no longer needed or is new.  This is simply
handled by deleting the old tag or creating a new one.

The relatively uncommon case is when the length of the tag changes.
This is only likely to happen when the implementation of the tagger has
changed (i.e., a new version of the contributing pluin) so that it now
stores different data.  In this case the old tag is deleted and a new
one is created.

----
A PDOM read lock is needed for all database reads, and a write lock is
needed for all writes (running without assertions masked the problem
until now).  The tag implementation needs to provide read access in
several places and write access only during tag creation.

The new read interface is accessible from the PDOM binding.  The
existing CDT implementation already ensures a read lock while the PDOM
binding is being accessed.

The new write interface is only accessible while the tags are being
copied from the in-memory values to the PDOM.  The existing CDT
implementation already ensures a write lock during this interval.

----
If an extension's class cannot be instantiated once, there is no reason
to continue trying.  This marks the invalid extension so only one
instantiation attempt is made.

----
Also fixes a bug where all tags in the index where returned instead of
only the tags for the specified record.

Change-Id: Ia5857a518587368417e9ecb08d5130d75b8fa0fa