Update patch set 2
Patch Set 2:

Here is an updated patch that fixes the infinite recursion in the way you suggested, and also fixes a regression I saw in my codebase (I added a new test for it). Please let me know if you see any other regressions in yours.

Regarding the addition of constructors taking an "IASTNode pointOfDefinition" parameter to evaluations other than EvalFixed: is there a way to do this without duplicating the constructor body of every evaluation?

Also, how did you tell that the infinite recursion happened in IndexCPPTemplateResolutionTest.testConstantPropagationFromHeader? When I ran that test, I saw an AssertionError because a name could not be found, but no StackOverflowError.

Patch-set: 2