Update patch set 2
Patch Set 2:

The regressions no longer occur with the latest patch.

The IASTName-based constructor can refer to the IBinding-based one. For example for EvalBinary it can be written as:

public EvalBinary(int operator, ICPPEvaluation arg1, ICPPEvaluation arg2, IASTNode pointOfDefinition) {
    this(operator, arg1, arg2, findEnclosingTemplate(pointOfDefinition));
}

This code assumes that findEnclosingTemplate method is moved to CPPDependentEvaluation class.

I saw infinite recursion in IndexCPPTemplateResolutionTest.testConstantPropagationFromHeader when I ran that test in debugger with a breakpoint on StackOverflowError. The stack overflow happened on the indexer thread, not on the main one.

Patch-set: 2